requirements.txt
matplotlib==3.10.6
numpy==2.3.3
prettytable==3.16.0
opencv_python==4.12.0.88
pandas
torch
joblib
tqdm
scipy
torch
kornia
Pillow
python_Levenshtein
requests
pyyaml
openai
